Some basic symmetry
Created a MapHelper class that is for handling MapLocations.

Also created a snippet that updates what quadrants a bot has been in during the setup phase. If we've been in certain quadrants, we can rule out horizontal/vertical symmetry (though never rotational -- we'll have to handle that after the walls fall). I think it's fairly safe to assume that if we can't exclude either horizontal or vertical symmetry, it is very likely that the map will actually be horizontally/vertically symmetric.